Introduce logging, remove memory copy
